autofill/name.html

<!DOCTYPE HTML>
<head>
<meta charset="utf-8" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/autofill-helper.js"></script>
<style>
  :autofill { color: red; }
</style>
</head>
<body>
<form id="test_form">
  <input type="text" autocomplete="name" id="test_field">
</form>
<script>
  promise_test(async t => {
    if (!window.test_driver) {
      return;
    }
    // Save form data to autofill.
    const form = create_form_for_data({"name": "foobar"});
    await window.test_driver.save_to_autofill(form);

    // Now test that the values were properly filled
    await test_driver.apply_autofill_suggestions(
      document.getElementById("test_field"));
    await wait_on_element_autofill();
    assert_element_has_value("test_field", "foobar");
  });
</script>
